fix(config): keep TLS validation below runtime boundary
